+++ core/components/romanesco/elements/plugins/07_computations/c_content/manipulatedom.plugin.php
@@ -36,11 +36,11 @@ switch ($modx->event->name) {
         }
 
         // Add header class to content headers without class name
-        $dom->filter('h1:not(.ui.header)')->addClass('ui header');
-        $dom->filter('h2:not(.ui.header)')->addClass('ui header');
-        $dom->filter('h3:not(.ui.header)')->addClass('ui header');
-        $dom->filter('h4:not(.ui.header)')->addClass('ui header');
-        $dom->filter('h5:not(.ui.header)')->addClass('ui header');
+        $dom->filter('h1:not(.header)')->addClass('ui header');
+        $dom->filter('h2:not(.header)')->addClass('ui header');
+        $dom->filter('h3:not(.header)')->addClass('ui header');
+        $dom->filter('h4:not(.header)')->addClass('ui header');
+        $dom->filter('h5:not(.header)')->addClass('ui header');
 
         // Inject inverted classes
         $dom->filter('.inverted.segment')